Add NavigationLinks constant for header navigation
